Developer Build 2015.0.2
Wrote the arm lift API. It will lift till it hits a limit switch and
then stop. Works for both up and down, and any amount of limit switches.
Assumes we put a limit switch at the starting point. Will not go lower
than intended, need to make it so it does not go higher than intended.